Added continuous pot -RC

+++ LRT14/ComponentData/LauncherLoaderData.h
@@ -20,7 +20,7 @@ public:
 	bool GetLoad();
 
 	bool IsLoadingComplete();
-	int GetSensorValue();
+	float GetSensorValue();
 	bool IsBallDetected();
 	
 	void ResetCommands();
@@ -28,14 +28,14 @@ public:
 	
 private:
 	void SetLoadingComplete(bool complete);
-	void SetSensorValue(int value);
+	void SetSensorValue(float value);
 	void SetBallDetected(bool detected);
 	
 	bool m_fire;
 	bool m_purge;
 	bool m_load;
 	bool m_complete;
-	int m_value;
+	float m_value;
 	bool m_ballDetected;
 	
 	friend class LauncherLoader;
